land: squash 3 attempt commits from afk/3853-decompose-terminal-ts-zero-slack-baselin
- refactor(dev): extract terminal envelope assembly
- refactor(dev): re-export terminal envelope assembly
- chore(dev): lower terminal size baseline
